requirements.txt
numpy==1.24.3
scipy==1.11.2
pyzmq==25.1.1
sounddevice==0.4.6
librosa==0.10.1
faster-whisper==1.0.0
networkx==3.2
flask==3.0.0
flask-socketio==5.3.5
plotly==5.18.0
pyyaml==6.0.1
pydantic==2.5.0
loguru==0.7.2
pytest==7.4.3
pytest-timeout==2.2.0
vaderSentiment>=3.3.2
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
piper-tts>=1.2.0
psutil>=5.9.0
